adding middleware on the endpoitns

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Http\Controllers\AccountController;
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\TransactionController;
 use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
@@ -17,32 +18,27 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-    echo 'hello there';
-    return $request->user();
-});
 
 
-Route::get('/accounts', [AccountController::class, 'index']);
-Route::post('/accounts', [AccountController::class, 'store']);
-Route::put('/accounts/{account}', [AccountController::class, 'update']);
-Route::delete('/accounts/{account}', [AccountController::class, 'destroy']);
+Route::get('/accounts', [AccountController::class, 'index'])->middleware('auth:sanctum');
+Route::post('/accounts', [AccountController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/accounts/{account}', [AccountController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/accounts/{account}', [AccountController::class, 'destroy'])->middleware('auth:sanctum');
 
-Route::get('/users', [UserController::class, 'index']);
-Route::post('/users', [UserController::class, 'store']);
-Route::put('/users/{user}', [UserController::class, 'update']);
-Route::delete('/users/{user}', [UserController::class, 'destroy']);
-
-
-Route::get('/transactions', [TransactionController::class, 'index']);
-Route::post('/transactions', [TransactionController::class, 'store']);
-Route::put('/transactions/revert/{journalId}', [TransactionController::class, 'revertTransaction']);
-Route::put('/transactions/{transaction}', [TransactionController::class, 'update']);
-Route::delete('/transactions/{transaction}', [TransactionController::class, 'destroy']);
+Route::get('/users', [UserController::class, 'index'])->middleware('auth:sanctum');
+Route::post('/users', [UserController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/users/{user}', [UserController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/users/{user}', [UserController::class, 'destroy'])->middleware('auth:sanctum');
 
 
-Route::get('/', function () {
-    return response()->json([
-        'message' => 'Welcome to our api',
-    ]);
-});
+Route::get('/transactions', [TransactionController::class, 'index'])->middleware('auth:sanctum');
+Route::post('/transactions', [TransactionController::class, 'store'])->middleware('auth:sanctum');
+Route::put('/transactions/revert/{journalId}', [TransactionController::class, 'revertTransaction'])->middleware('auth:sanctum');
+Route::put('/transactions/{transaction}', [TransactionController::class, 'update'])->middleware('auth:sanctum');
+Route::delete('/transactions/{transaction}', [TransactionController::class, 'destroy'])->middleware('auth:sanctum');
+
+
+
+Route::post('/auth/register', [AuthController::class, 'createUser']);
+Route::post('/auth/login', [AuthController::class, 'loginUser']);
+
